+++ examples/v2.0.1/views/Index.js
@@ -7,7 +7,7 @@ import { push } from 'react-router-redux';
 // import { TweenMax } from 'gsap';// 动画
 // 
 // var request = require('request');
-import { push as basenamePush } from '../../lib_modules/router-basename';
+import { push as basenamePush } from '../../../lib_modules/router-basename';
 
 class IndexView extends Component{
   static contextTypes={
@@ -212,7 +212,9 @@ class IndexView extends Component{
   }
 }
 
-function select(store){
+function select(store/*, ownProps*/){ // 1）第一个参数总是state对象，还可以使用第二个参数，代表容器组件的props对象
+								  // 2) 侦听 Store，每当state更新的时候，就会自动执行，重新计算 UI 组件的参数，从而触发 UI 组件的重新渲染。
+								  // 3）当使用了 ownProps 作为参数后，如果容器组件的参数发生变化，也会引发 UI 组件重新渲染。
 	return {
     basename: store.config.basename,
     detailState: store.router.location&&store.router.location.detailState