+++ lib_modules/router-basename/push.js
@@ -1,0 +1,9 @@
+const push = (loc) => {
+  let url = loc.props.basename + loc.pathname;
+  loc.pathname = url;
+  return (history)=>{
+    history? history.push(loc): loc.props.history.push(loc);
+  }
+}
+
+export default push