+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 ### 路由
 
 * BrowserRouter 路由封装 basename={} 处理
+* 将分离的模块，重复代码进行再次抽取
+* antdesign 组件模块的 `<style type="text/css"` 更新为 `<link` 引用
 
 ## 2.0.0 (2017年09月02日)